major refactor
me when i refactor every line of code in a single commit
