+++ config.env.example
@@ -1,8 +1,19 @@
+# Optionally, set timezone for logs
+#TIMEZONE=""
+
+# Set operating mode; defaults to relay
+# 'reset' mode will erase all ratings on ListenBrainz and/or LastFM.
+# MODE=""
+
+# Uncomment to enable two-way sync
+# (sync ratings from LastFM/Listenbrainz to Plex)
+# TWO_WAY="true"
+
 # Full Plex server URL including port, i.e. http://192.168.0.1:32400
-SERVER_URL=
+PLEX_SERVER_URL=""
 
 # Title of the library where music is stored
-MUSIC_LIBRARY=Music
+MUSIC_LIBRARY="Music"
 
 # Threshold to determine a "Loved" track.
 # 1 star = 2; 2 stars = 4, etc.
@@ -17,18 +28,18 @@ LOVE_THRESHOLD=10
 #HATE_THRESHOLD=2
 
 # Obtain these at https://www.last.fm/api/account/create
-LASTFM_API_KEY=
-LASTFM_SECRET=
-LASTFM_USERNAME=
-LASTFM_PASSWORD=
+LASTFM_TOKEN=""  # API key
+LASTFM_SECRET=""
+LASTFM_USERNAME=""
+LASTFM_PASSWORD=""
 
 # ListenBrainz username and token - leave blank if you do not wish to use ListenBrainz
 # Obtain token at: https://listenbrainz.org/settings/
-LISTENBRAINZ_TOKEN=
-LISTENBRAINZ_USERNAME=
+LISTENBRAINZ_TOKEN=""
+LISTENBRAINZ_USERNAME=""
 
 # Databse file
-DATABASE=ratingrelay.db
+DATABASE="ratingrelay.db"
 
 # You do not need to set these.
 PLEX_CID=
