+++ ratingrelay/__init__.py
@@ -1,0 +1,3 @@
+from .ratingrelay import main
+
+__all__ = ["main"]
