+++ ratingrelay/exceptions.py
@@ -1,0 +1,6 @@
+class ConfigError(Exception):
+    """Raised when required configuration value is missing or invalid"""
+
+
+class LibraryNotFoundError(Exception):
+    """Raised when a matching music library is not found on Plex server"""
