+++ ratingrelay/services.py
@@ -1,0 +1,14 @@
+from dataclasses import dataclass
+from typing import Optional
+from .plex import Plex
+from .database import Database
+from .lastfm import LastFM
+from .listenbrainz import ListenBrainz
+
+
+@dataclass
+class Services:
+    plex: Plex
+    db: Database
+    lfm: Optional[LastFM]
+    lbz: Optional[ListenBrainz]
